requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asteval==1.0.6
astropy==7.1.0
astropy-iers-data==0.2025.9.1.0.42.11
astroquery==0.4.10
beautifulsoup4==4.13.5
blosc2==3.7.2
caugetch==0.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
clipboard==0.0.4
colorama==0.4.6
configparser==7.2.0
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
dill==0.4.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
flake8==7.3.0
fonttools==4.59.2
getpass4==0.0.14.1
h11==0.16.0
h5py==3.14.0
habanero==2.3.0
hitran-api==1.2.2.4
hjson==3.1.0
html5lib==1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
Jinja2==3.1.6
joblib==1.5.2
json-tricks==3.17.3
keyring==25.6.0
kiwisolver==1.4.9
llvmlite==0.44.0
lmfit==1.3.4
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
mpldatacursor==0.7.1
msgpack==1.1.1
narwhals==2.3.0
ndindex==1.10.0
numba==0.61.2
numexpr==2.11.0
numpy==2.2.6
orjson==3.11.2
packaging==25.0
pandas==2.3.2
PeakUtils==1.3.5
periodictable==2.0.2
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
pluggy==1.6.0
psutil==7.0.0
publib==0.4.0
py-cpuinfo==9.0.0
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyerfa==2.0.1.5
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
pyperclip==1.9.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyvo==1.7
PyYAML==6.0.2
redis==6.4.0
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scipy==1.16.1
seaborn==0.13.2
SecretStorage==3.3.3
sentry-sdk==2.35.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.47.2
synthetick==0.1.0
tables==3.10.2
termcolor==3.1.0
toml==0.10.2
tqdm==4.67.1
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
uncertainties==3.2.3
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vulkan==1.3.275.1
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
